cli/cmd: move deleteHelmRelease to pkg/components/util
As component-oriented functionality should be stored in pkg/components
and cli package should not go too much into helm details etc.
